featured_plugins/dns_resolver/requirements.txt
mmh3
dnspython